+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+#### 11.0.0
+
+- Added a `isBase32()` validator
+  ([#1023](https://github.com/chriso/validator.js/pull/1023))
+- Updated `isEmail()` to validate display names according to RFC2822
+  ([#1004](https://github.com/chriso/validator.js/pull/1004))
+- Updated `isEmail()` to check total email length
+  ([#1007](https://github.com/chriso/validator.js/pull/1007))
+- The internal `toString()` util is no longer exported
+  ([0277eb](https://github.com/chriso/validator.js/commit/0277eb00d245a3479af52adf7d927d4036895650))
+- New and improved locales
+  ([#999](https://github.com/chriso/validator.js/pull/999),
+   [#1010](https://github.com/chriso/validator.js/pull/1010),
+   [#1017](https://github.com/chriso/validator.js/pull/1017),
+   [#1022](https://github.com/chriso/validator.js/pull/1022),
+   [#1031](https://github.com/chriso/validator.js/pull/1031),
+   [#1032](https://github.com/chriso/validator.js/pull/1032))
+
 #### 10.11.0
 
 - Fix imports like `import .. from "validator/lib/.."`